refactor(ritobin): move the visitor tests into their own file
